Base bloc call ddang loi
