+++ pubspec.yaml
@@ -66,10 +66,12 @@ dependencies:
   flutter_launcher_icons: ^0.14.3
   #video call:
   stringee_plugin: ^1.2.14
+  uuid: ^4.2.1
   provider: ^6.0.5
   flutter_local_notifications: ^18.0.1
   dart_jsonwebtoken: ^2.17.0
   agora_rtc_engine: ^6.5.1
+
   #test:
   mockito: ^5.4.3
   bloc_test: ^10.0.0
